Network Diffusion using

Network Diffusion using

Usage

ND_s(
  X0,
  W,
  alpha = 0.7,
  nMax = 10000,
  eps = 1e-06,
  final.smooth = FALSE,
  all.steps = FALSE,
  verbose = FALSE
)

Arguments

X0

matrix composed of column vectors with initial distribution of information

W

symmetrically normalized adjacency matrix W = D^-1 A D^-1, see 'normalize_adj_mat' function

alpha

numeric, the smothing factor

nMax

numeric, maximum number of iterations

eps

numeric, the iteration will stop when the maximum difference between matrix Xs between two consecutive iteraction is smaller than eps

final.smooth

TRUE/FALSE, whether to do the final step of smoothing

all.steps,

TRUE/FALSE, whether to store all steps

verbose,

TRUE/FALSE

Value

a list with:

Xt the smoothed matrix;

eps see above;

max.abs.diff max(abs(F_t) - abs(F_t-1));

Xs.all transient Xs matrices.