Network diffusion

Network diffusion: input scores (X0) are smoothed by network-diffusion, obtaining the corresponding network-constrained scores Xs.

Usage

run_ND(
  X0 = NULL,
  W = NULL,
  alpha = 0.7,
  nMax = 10000,
  eps = 1e-06,
  BPPARAM = NULL
)

Arguments

X0

matrix or a list of matrices (if X0 was permuted (see 'perm_X0' function), where the first element of the list is the one obtained with real data); each column (layer) of the matrix X0 is a score vector over all vertices of G.

W

matrix; symmetrically normalized adjacency matrix W = D^-1 A D^-1, see 'normalize_adj_mat' function

alpha

numeric; the smothing factor

nMax

numeric; maximum number of iterations

eps

numeric; the iteration will stop when the maximum difference between matrix Xs between two consecutive iteraction is smaller than eps

BPPARAM

optional BiocParallelParam instance determining the parallel back-end to be used during evaluation. If NULL, parallel evaluation is disabled using SerialParam(). See ?bplapply.

Value

a matrix or a list of matrices (if data were permuted) with network diffusion scores.